Extract Neutral class to separate module
